Add n teammates to a feedback session
A feedback session will have more than 2 teammates on them so there
needs to be a way to add many more teammates, in fact, there needs to be
a way to add n teammates.

I accomplish this via a stimulus controller to append new html to the
bottom of the document.

To test this I created the first system test. It uses selenium as it's
default driver.