Create the "Feedback Lineup" screen
After a session gets created the next step is to create the "feedback
lineup". These are the groups of 2 people (one person giving feedback
and the other person receiving it).

This commit creates the simplest of feedback lineups. There's no logic
going into creating the groups. All I'm doing right now is creating the
lineup and displaying on the screen.