Generate and show feedback rotations
At this point in the process we have an approved feedback lineup. The
next step is to arrange the lineup in a way that can be most
efficiently done (i.e. if someone is receiving feedback they can't be
giving it at the same time). This can be done manually but having done
via the software is a nice-ity (as I've found).

The algorithm to compute the feedback rotations is really bad right now.
But it's tested and works so I'm moving forward for now.