Fix bug in lineup manager grouping
The bug I was seeing had a person giving feedback to themselves.

Groupings need to based on the lineup from the protocol not the original
list of passed in teammates. With the addition of the random protocol
the lineup from the protocol will be different from the passed in
teammates array.